[NEW] moved .json files && file access corrections
