+++ src/jsons/personalities.json
